backend/requirements.txt
celery>=5.3
Django>=5.0,<6.0
djangorestframework>=3.15
gunicorn>=21.2
psycopg[binary]>=3.1
python-dotenv>=1.0
djangorestframework-simplejwt
django-cors-headers
drf-spectacular
dj-database-url
reportlab
django-redis
redis
sentry-sdk
cryptography>=42.0.0
pymupdf
pyotp>=2.9.0
django-storages>=1.14.4
boto3>=1.34.0
qrcode>=8.0